fix: reject blank s3 startup configuration
